server/requirements.txt
django-admin
django
djangorestframework